# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationOut": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "changeVerdicts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Changeverdicts",
            "type": "array"
          },
          "displayState": {
            "title": "Displaystate",
            "type": "string"
          },
          "hasInvoiceOverprice": {
            "title": "Hasinvoiceoverprice",
            "type": "boolean"
          },
          "hasUnjustifiedIncrease": {
            "title": "Hasunjustifiedincrease",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isAssessable": {
            "title": "Isassessable",
            "type": "boolean"
          },
          "modelLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modellabel"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "$ref": "#/components/schemas/EvaluationStatus"
          }
        },
        "required": [
          "id",
          "status",
          "attempt",
          "changeVerdicts",
          "hasUnjustifiedIncrease",
          "hasInvoiceOverprice",
          "isAssessable",
          "note",
          "modelLabel",
          "displayState"
        ],
        "title": "EvaluationOut",
        "type": "object"
      },
      "EvaluationStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "needs_review",
          "failed"
        ],
        "title": "EvaluationStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolveAppealInput": {
        "additionalProperties": false,
        "properties": {
          "newVerdicts": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newverdicts"
          },
          "note": {
            "maxLength": 600,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "ResolveAppealInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifySnapshotInput": {
        "additionalProperties": false,
        "properties": {
          "approve": {
            "title": "Approve",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "approve"
        ],
        "title": "VerifySnapshotInput",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API نمایشی «ترب تعمیر». دادهٔ نمونه، پرداخت شبیه‌سازی‌شده و پاسخ AI نمایشی در این نصب صریحاً برچسب‌گذاری می‌شوند.",
    "title": "ترب تعمیر — API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/support/appeals/{appeal_id}/resolve": {
      "post": {
        "description": "Correcting the evidence or re-running the judgement.\n\nA manual change keeps the previous result on record and always names a reason and the\nacting person. Support never overrides an applied settlement.",
        "operationId": "resolve_appeal_api_support_appeals__appeal_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "appeal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appeal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAppealInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resolve Appeal Api Support Appeals  Appeal Id  Resolve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Appeal",
        "tags": [
          "support"
        ]
      }
    },
    "/api/support/evaluations/{request_id}": {
      "get": {
        "operationId": "get_evaluation_api_support_evaluations__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EvaluationOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Evaluation Api Support Evaluations  Request Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "support"
        ]
      }
    },
...
    "/api/support/price-snapshots/{snapshot_id}/verify": {
      "post": {
        "description": "One party typing a number is never enough to dock a score.",
        "operationId": "verify_snapshot_api_support_price_snapshots__snapshot_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifySnapshotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verify Snapshot Api Support Price Snapshots  Snapshot Id  Verify Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Snapshot",
        "tags": [
          "support"
        ]
      }
    },
...
    "/api/support/requests/{request_id}/evaluation": {
      "get": {
        "operationId": "request_evaluation_api_support_requests__request_id__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Evaluation",
        "tags": [
          "support"
        ]
      }
    }
  }
}